Windows, JNI: move process mgmt to process.cc
Move process lifetime management functions to
process.cc and process.h, into the WaitableProcess
utility class.

The next step is moving the
WaitableProcess-related state from NativeProcess
into WaitableProcess, then to use WaitableProcess
in the Windows-native test wrapper.
